Move webcompat tests to TS schema
