+++ tests/webcompat-tests.js
@@ -1,9 +1,7 @@
 const expect = require('chai').expect
 const fs = require('fs')
-const Ajv = require('ajv').default
-const ajv = new Ajv()
-const rootSchema = JSON.parse(fs.readFileSync('./tests/schemas/webcompat-settings.json', 'utf8'))
-const validateRoot = ajv.compile(rootSchema)
+const { createValidator, formatErrors } = require('./schema-validation')
+const validateRoot = createValidator('WebCompatSettings')
 
 const platforms = ['macos', 'ios', 'android']
 const latestConfigs = platforms.map((plat) => {
@@ -17,12 +15,7 @@ for (const config of latestConfigs) {
     describe(`validates the config settings for ${config.name}`, () => {
         it('should have valid webCompat setting', () => {
             const actual = validateRoot(config.body.features.webCompat.settings)
-            if (validateRoot.errors) {
-                for (const error of validateRoot.errors) {
-                    console.error(error)
-                }
-            }
-            expect(actual).to.be.equal(true)
+            expect(actual).to.be.equal(true, formatErrors(validateRoot.errors))
         })
     })
 }
